Removed compiler warning (gcc4.4 x86_64)
git-svn-id: http://svn.ta3d.org/yuni/trunk@831 7e728022-c0ab-4603-8e65-c666ead28a33

+++ src/yuni/core/getopt/option.cpp
@@ -112,7 +112,7 @@ namespace GetOptImpl
 				out << "\n                             ";
 			else
 			{
-				for (int i = 6 + 2 + 1 + longName.size() + (requireParameter ? 6 : 0); i < 30; ++i)
+				for (unsigned int i = 6 + 2 + 1 + (unsigned int) longName.size() + (requireParameter ? 6 : 0); i < 30; ++i)
 					out.put(' ');
 			}
 		}
